requirements.txt
# VisualVoice — Production Requirements
# ----------------------------------------
# Pin major versions for reproducibility.
# Run: pip install -r requirements.txt

# ── Core Streamlit ─────────────────────
streamlit>=1.32.0,<2.0.0

# ── AI / ML ────────────────────────────
transformers>=4.38.0,<5.0.0
torch>=2.2.0                     # CPU-only is fine; use torch+cu121 for GPU
Pillow>=10.2.0,<11.0.0           # Image I/O for BLIP pipeline
accelerate>=0.27.0               # Optional: speeds up transformers inference

# ── Grok / OpenAI-compatible client ───
openai>=1.12.0,<2.0.0

# ── HTTP & env ─────────────────────────
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0

# ── Utilities ──────────────────────────
# (all stdlib on Python 3.9+; listed for clarity)
# pathlib, logging, zipfile, io, uuid — stdlib, no install needed